fixed errors

+++ app.rb
@@ -10,7 +10,7 @@ get("/square/new") do
 end
 
 get("/square/results") do
-  @number = params.fetch("number").to_i
+  @number = params.fetch("number").to_f
   erb (:square_results)
 end
 
@@ -19,7 +19,7 @@ get("/square_root/new") do
 end
 
 get("/square_root/results") do
-  @user_number = params.fetch("user_number").to_i
+  @user_number = params.fetch("user_number").to_f
   erb(:square_root_results)
 end
